requirements.txt
streamlit
python-dotenv
pickle-mixin
langchain
faiss-cpu
unstructured
openai
gdown